refactor: refactor movies index
