+++ lib/media_server/movies_index.ex
@@ -39,49 +39,15 @@ defmodule MediaServer.MoviesIndex do
     movie["title"]
   end
 
-  def some_value({:ok, value}) do
-    value
-  end
-
-  def some_value(:error) do
-    ""
-  end
-
-  def another_test(nil, _url_type) do
-    ""
-  end
-
-  def another_test(value, url_type) do
-    Map.fetch(value, url_type)
-    |> some_value()
-  end
-
-  def some_test({:ok, value}, "headshot") do
-    Enum.find(value, fn item -> item["coverType"] === "headshot" end)
-    |> another_test("url")
-  end
-
-  def some_test({:ok, value}, type) do
-    Enum.find(value, fn item -> item["coverType"] === type end)
-    |> another_test("remoteUrl")
-  end
-
-  def some_test(:error, _type) do
-    ""
-  end
-
   def get_poster(movie) do
-    Map.fetch(movie, "images")
-    |> some_test("poster")
+    MediaServer.Helpers.get_poster(movie)
   end
 
   def get_background(movie) do
-    Map.fetch(movie, "images")
-    |> some_test("fanart")
+    MediaServer.Helpers.get_background(movie)
   end
 
   def get_headshot(movie) do
-    Map.fetch(movie, "images")
-    |> some_test("headshot")
+    MediaServer.Helpers.get_headshot(movie)
   end
 end
