+++ test/media_server/helpers_test.exs
@@ -1,7 +1,7 @@
-defmodule MediaServer.MoviesIndexTest do
+defmodule MediaServer.HelpersTest do
   use ExUnit.Case
 
-  @some_movie %{
+  @some_media %{
     "images" => [
       %{
         "coverType" => "poster",
@@ -18,7 +18,7 @@ defmodule MediaServer.MoviesIndexTest do
     ]
   }
 
-  @some_movie_without_poster %{
+  @some_media_without_poster %{
     "images" => [
       %{
         "coverType" => "fanart",
@@ -31,7 +31,7 @@ defmodule MediaServer.MoviesIndexTest do
     ]
   }
 
-  @some_movie_without_fanart %{
+  @some_media_without_fanart %{
     "images" => [
       %{
         "coverType" => "poster",
@@ -44,7 +44,7 @@ defmodule MediaServer.MoviesIndexTest do
     ]
   }
 
-  @some_movie_without_headshot %{
+  @some_media_without_headshot %{
     "images" => [
       %{
         "coverType" => "poster",
@@ -58,30 +58,30 @@ defmodule MediaServer.MoviesIndexTest do
   }
 
   test "it should get empty string without images" do
-    assert MediaServer.MoviesIndex.get_poster(%{}) === ""
+    assert MediaServer.Helpers.get_poster(%{}) === ""
   end
 
   test "it should get poster" do
-    assert MediaServer.MoviesIndex.get_poster(@some_movie) === "https://some.remote.url/some-poster"
+    assert MediaServer.Helpers.get_poster(@some_media) === "https://some.remote.url/some-poster"
   end
 
   test "it should get empty string without poster" do
-    assert MediaServer.MoviesIndex.get_poster(@some_movie_without_poster) === ""
+    assert MediaServer.Helpers.get_poster(@some_media_without_poster) === ""
   end
 
   test "it should get background" do
-    assert MediaServer.MoviesIndex.get_background(@some_movie) === "https://some.remote.url/some-fanart"
+    assert MediaServer.Helpers.get_background(@some_media) === "https://some.remote.url/some-fanart"
   end
 
   test "it should get empty string without background" do
-    assert MediaServer.MoviesIndex.get_background(@some_movie_without_fanart) === ""
+    assert MediaServer.Helpers.get_background(@some_media_without_fanart) === ""
   end
 
   test "it should get headshot" do
-    assert MediaServer.MoviesIndex.get_headshot(@some_movie) === "https://some.remote.url/some-headshot"
+    assert MediaServer.Helpers.get_headshot(@some_media) === "https://some.remote.url/some-headshot"
   end
 
   test "it should get empty string without headshot" do
-    assert MediaServer.MoviesIndex.get_headshot(@some_movie_without_headshot) === ""
+    assert MediaServer.Helpers.get_headshot(@some_media_without_headshot) === ""
   end
 end
